requirements.txt
absl-py==2.1.0
agora-token-builder==1.0.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.0.1
av==12.3.0
blinker==1.8.2
blis==1.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.10
datasets==3.0.2
debugpy==1.8.9
decorator==5.1.1
dill==0.3.8
distlib==0.3.7
distro==1.9.0
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-restframework==0.0.1
djangorestframework-simplejwt==5.3.0
dnspython==2.7.0
EasyProcess==1.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoint2==1.1
errant==3.0.0
executing==2.1.0
ffmpeg-python==0.2.0
filelock==3.12.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
google-pasta==0.2.0
gramformer @ git+https://github.com/PrithivirajDamodaran/Gramformer.git@23425cd2e98a919384cab6156af8adf1c9d0639a
grpcio==1.66.1
h11==0.14.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2024.10.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
hyperframe==5.2.0
idna==3.10
ifaddr==0.2.0
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.26.1
libclang==18.1.1
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
namex==0.0.8
narwhals==1.9.3
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==0.28.0
openai-whisper @ git+https://github.com/openai/whisper.git@173ff7dd1d9fb1c4fddea0d41d704cfefeb8908c
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
phonenumbers==8.13.24
pillow==10.4.0
platformdirs==4.3.2
preshed==3.0.9
prettytable==3.12.0
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
PyAutoGUI==0.9.54
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.8.0
pylibsrtp==0.10.0
PyMsgBox==1.0.9
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
python-slugify==8.0.4
pytweening==1.0.7
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
redis==5.0.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
sentence-transformers==3.2.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.11.0
sqlparse==0.5.1
srsly==2.5.0
stack-data==0.6.3
streamlit==1.39.0
streamlit-webrtc==0.47.9
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.4.0
text-unidecode==1.3
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.24.5
wasabi==1.1.3
watchdog==5.0.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.16.0
xgboost==2.1.1
xxhash==3.5.0
yarl==1.17.1
yellowbrick==1.5
